Fixed float typing consistency

+++ experiments/inpainting_experiment.py
@@ -61,7 +61,7 @@ def compute_metrics(dataset_path, metric_list, sampler, n_samples, n_steps,
 
     with open(dataset_path, "rb") as f:
         ref_data = pkl.load(f)
-    X_ref = from_numpy(ref_data["X"])
+    X_ref = from_numpy(ref_data["X"]).to(torch.float32)
 
     metric_results = {}
     for metric in metric_list:
